Install pkg-config files to root multiarch dir.
The ogre-next-2.2 package had a loop in debian/rules to rename and
install pkg-config files after cmake installation.

When porting those changes to quilt patches I renamed the files but did
not relocate them. This patch builds on the previous versioned pkgconfig
filenames and installs them in the multiarch lib's top-level pkgconfig
directory which should be on the system's default `PKG_CONFIG_PATH`.